+++ Exception/SortAttributeNotAvailableException.php
@@ -6,15 +6,14 @@ namespace Oka\PaginationBundle\Exception;
  * @author Cedrick Oka Baidai <okacedrick@gmail.com>
  * 
  */
-class SortAttributeNotAvailableException extends \Exception
+class SortAttributeNotAvailableException extends PaginationException
 {
 	/**
 	 * @var string $sort
 	 */
 	private $sort;
 	
-	public function __construct($sort, $message = null, $code = null, $previous = null)
-	{
+	public function __construct($sort, $message = null, $code = null, $previous = null) {
 		parent::__construct($message, $code, $previous);
 		
 		$this->sort = $sort;
@@ -23,8 +22,7 @@ class SortAttributeNotAvailableException extends \Exception
 	/**
 	 * @return string
 	 */
-	public function getSort()
-	{
+	public function getSort() {
 		return $this->sort;
 	}
 }